Split arduino wrapper into SATS/arduino.sats

+++ serial_ats/Arduino.h
@@ -97,7 +97,7 @@ void init(void);
 
 void pinMode(uint8_t, uint8_t);
 void digitalWrite(uint8_t, uint8_t);
-int digitalRead(uint8_t);
+uint8_t digitalRead(uint8_t);
 int analogRead(uint8_t);
 void analogReference(uint8_t mode);
 void analogWrite(uint8_t, int);
